dev_requirements.txt
South==0.8.4
MySQL-python==1.2.5
django_compressor==1.3
pytz>2012